+++ src/Adapters/Contracts/TagNormalizerAware.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Cosmastech\StatsDClient\Adapters\Contracts;
+
+use Cosmastech\StatsDClient\TagNormalizers\TagNormalizer;
+
+interface TagNormalizerAware
+{
+    public function setTagNormalizer(TagNormalizer $tagNormalizer): void;
+}
