+++ src/Adapters/InMemory/InMemoryClientAdapter.php
@@ -1,0 +1,132 @@
+<?php
+
+namespace Cosmastech\StatsDClient\Adapters\InMemory;
+
+use Cosmastech\StatsDClient\Adapters\Concerns\TagNormalizerAwareTrait;
+use Cosmastech\StatsDClient\Adapters\Contracts\TagNormalizerAware;
+use Cosmastech\StatsDClient\Adapters\InMemory\Models\InMemoryCountRecord;
+use Cosmastech\StatsDClient\Adapters\InMemory\Models\InMemoryDistributionRecord;
+use Cosmastech\StatsDClient\Adapters\InMemory\Models\InMemoryGaugeRecord;
+use Cosmastech\StatsDClient\Adapters\InMemory\Models\InMemoryHistogramRecord;
+use Cosmastech\StatsDClient\Adapters\InMemory\Models\InMemorySetRecord;
+use Cosmastech\StatsDClient\Adapters\InMemory\Models\InMemoryStatsRecord;
+use Cosmastech\StatsDClient\Adapters\InMemory\Models\InMemoryTimingRecord;
+use Cosmastech\StatsDClient\Adapters\StatsDClientAdapter;
+use Cosmastech\StatsDClient\TagNormalizers\NoopTagNormalizer;
+use Psr\Clock\ClockInterface;
+
+class InMemoryClientAdapter implements StatsDClientAdapter, TagNormalizerAware
+{
+    use TagNormalizerAwareTrait;
+
+    protected InMemoryStatsRecord $stats;
+    protected readonly ClockInterface $clock;
+
+    public function __construct(ClockInterface $clock)
+    {
+        $this->clock = $clock;
+
+        $this->reset();
+        $this->setTagNormalizer(new NoopTagNormalizer());
+    }
+
+    public function timing(string $stat, float $durationMs, float $sampleRate = 1.0, array $tags = []): void
+    {
+        $this->stats->timing[] = new InMemoryTimingRecord(
+            $stat,
+            $durationMs,
+            $sampleRate,
+            $this->normalizeTags($tags),
+            $this->clock->now()
+        );
+    }
+
+    public function gauge(string $stat, float $value, float $sampleRate = 1.0, array $tags = []): void
+    {
+        $this->stats->gauge[] = new InMemoryGaugeRecord(
+            $stat,
+            $value,
+            $sampleRate,
+            $this->normalizeTags($tags),
+            $this->clock->now()
+        );
+    }
+
+    public function histogram(string $stat, float $value, float $sampleRate = 1.0, array $tags = []): void
+    {
+        $this->stats->histogram[] = new InMemoryHistogramRecord(
+            $stat,
+            $value,
+            $sampleRate,
+            $this->normalizeTags($tags),
+            $this->clock->now()
+        );
+    }
+
+    public function distribution(string $stat, float $value, float $sampleRate = 1.0, array $tags = []): void
+    {
+        $this->stats->distribution[] = new InMemoryDistributionRecord(
+            $stat,
+            $value,
+            $sampleRate,
+            $this->normalizeTags($tags),
+            $this->clock->now()
+        );
+    }
+
+    public function set(string $stat, float|string $value, float $sampleRate = 1.0, array $tags = []): void
+    {
+        $this->stats->set[] = new InMemorySetRecord(
+            $stat,
+            $value,
+            $sampleRate,
+            $this->normalizeTags($tags),
+            $this->clock->now()
+        );
+    }
+
+    public function increment(array|string $stats, float $sampleRate = 1.0, array $tags = [], int $value = 1): void
+    {
+        $this->updateStats($stats, $value, $sampleRate, $tags);
+    }
+
+    public function decrement(array|string $stats, float $sampleRate = 1.0, array $tags = [], int $value = -1): void
+    {
+        if ($value > 0) {
+            $value *= -1;
+        }
+
+        $this->updateStats($stats, $value, $sampleRate, $tags);
+    }
+
+    public function updateStats(array|string $stats, int $delta = 1, $sampleRate = 1.0, $tags = null): void
+    {
+        $stats = (array) $stats;
+        $now = $this->clock->now();
+
+        foreach ($stats as $stat) {
+            $this->stats->count[] = new InMemoryCountRecord(
+                $stat,
+                $delta,
+                $sampleRate,
+                $this->normalizeTags($tags),
+                $now
+            );
+        }
+    }
+
+    public function getStats(): InMemoryStatsRecord
+    {
+        return $this->stats;
+    }
+
+    public function reset(): void
+    {
+        $this->stats = new InMemoryStatsRecord();
+    }
+
+    public function getClient(): null
+    {
+        return null;
+    }
+}
