+++ src/Adapters/InMemory/Models/InMemoryCountRecord.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Cosmastech\StatsDClient\Adapters\InMemory\Models;
+
+use DateTimeImmutable;
+
+class InMemoryCountRecord
+{
+    public function __construct(
+        public string $stat,
+        public int $count,
+        public float $sampleRate,
+        public array $tags,
+        public DateTimeImmutable $recordedAt,
+    ) {
+    }
+}
