+++ src/Adapters/InMemory/Models/InMemoryDistributionRecord.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Cosmastech\StatsDClient\Adapters\InMemory\Models;
+
+use DateTimeImmutable;
+
+readonly class InMemoryDistributionRecord
+{
+    public function __construct(
+        public string $stat,
+        public float $value,
+        public float $sampleRate,
+        public array $tags,
+        public DateTimeImmutable $recordedAt,
+    ) {
+    }
+}
