+++ src/Adapters/InMemory/Models/InMemoryTimingRecord.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Cosmastech\StatsDClient\Adapters\InMemory\Models;
+
+use DateTimeImmutable;
+
+readonly class InMemoryTimingRecord
+{
+    public function __construct(
+        public string $stat,
+        public float $durationMilliseconds,
+        public float $sampleRate,
+        public array $tags,
+        public DateTimeImmutable $recordedAt,
+    ) {
+    }
+}
