--- src/Clients/Concerns/TagNormalizerAwareTrait.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Cosmastech\StatsDClient\Clients\Concerns;
-
-use Cosmastech\StatsDClient\TagNormalizers\TagNormalizer;
-
-trait TagNormalizerAwareTrait
-{
-    protected TagNormalizer $tagNormalizer;
-
-    public function setTagNormalizer(TagNormalizer $tagNormalizer): void
-    {
-        $this->tagNormalizer = $tagNormalizer;
-    }
-
-    protected function normalizeTags(array $tags): array
-    {
-        return $this->tagNormalizer->normalize($tags);
-    }
-}
