--- src/Clients/Contracts/TagNormalizerAware.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Cosmastech\StatsDClient\Clients\Contracts;
-
-use Cosmastech\StatsDClient\TagNormalizers\TagNormalizer;
-
-interface TagNormalizerAware
-{
-    public function setTagNormalizer(TagNormalizer $tagNormalizer): void;
-}
