+++ src/Clients/Datadog/DatadogLoggingClient.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Cosmastech\StatsDClient\Clients\Datadog;
+
+use DataDog\DogStatsd;
+use Psr\Log\LoggerInterface;
+use Psr\Log\LogLevel;
+
+class DatadogLoggingClient extends DogStatsd
+{
+    public function __construct(
+        protected readonly LoggerInterface $logger,
+        protected readonly string $logLevel = LogLevel::DEBUG,
+        array $datadogConfig = [],
+    ) {
+        parent::__construct($datadogConfig);
+    }
+
+    public function flush($message)
+    {
+        $this->logger->log($this->logLevel, $message);
+    }
+}
