+++ tests/Adapters/InMemory/InMemoryGaugeTest.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace Cosmastech\StatsDClient\Tests\Adapters\InMemory;
+
+use Cosmastech\StatsDClient\Adapters\InMemory\InMemoryClientAdapter;
+use Cosmastech\StatsDClient\Tests\BaseTestCase;
+use Cosmastech\StatsDClient\Tests\Doubles\ClockStub;
+use Cosmastech\StatsDClient\Tests\Doubles\TagNormalizerSpy;
+use DateTimeImmutable;
+use PHPUnit\Framework\Attributes\Test;
+
+class InMemoryGaugeTest extends BaseTestCase
+{
+    #[Test]
+    public function storesGaugeRecord()
+    {
+        // Given
+        $stubDateTime = new DateTimeImmutable("2018-02-13 18:50:00");
+        $inMemoryClient = new InMemoryClientAdapter(new ClockStub($stubDateTime));
+
+        // When
+        $inMemoryClient->gauge("gauge-stat", 23488);
+
+        // Then
+        $statsRecord = $inMemoryClient->getStats();
+        self::assertCount(1, $statsRecord->gauge);
+
+        $gaugeRecord = $statsRecord->gauge[0];
+        self::assertEquals("gauge-stat", $gaugeRecord->stat);
+        self::assertEquals(23488, $gaugeRecord->value);
+        self::assertEquals(1, $gaugeRecord->sampleRate);
+        self::assertEqualsCanonicalizing([], $gaugeRecord->tags);
+        self::assertEquals($stubDateTime, $gaugeRecord->recordedAt);
+    }
+
+    #[Test]
+    public function normalizesTags()
+    {
+        // Given
+        $inMemoryClient = new InMemoryClientAdapter(new ClockStub(new DateTimeImmutable()));
+
+        // And
+        $tagNormalizerSpy = new TagNormalizerSpy();
+        $inMemoryClient->setTagNormalizer($tagNormalizerSpy);
+
+        // When
+        $inMemoryClient->gauge(stat: "irrelevant", value: 1.0, tags: ["hello" => "world"]);
+
+        // Then
+        $this->assertEqualsCanonicalizing([["hello" => "world"]], $tagNormalizerSpy->getNormalizeCalls());
+    }
+}
