+++ tests/Adapters/InMemory/InMemoryHistogramTest.php
@@ -1,0 +1,57 @@
+<?php
+
+namespace Cosmastech\StatsDClient\Tests\Adapters\InMemory;
+
+use Cosmastech\StatsDClient\Adapters\InMemory\InMemoryClientAdapter;
+use Cosmastech\StatsDClient\Tests\BaseTestCase;
+use Cosmastech\StatsDClient\Tests\Doubles\ClockStub;
+use Cosmastech\StatsDClient\Tests\Doubles\TagNormalizerSpy;
+use DateTimeImmutable;
+use PHPUnit\Framework\Attributes\Test;
+
+class InMemoryHistogramTest extends BaseTestCase
+{
+    #[Test]
+    public function storesHistogramRecord()
+    {
+        // Given
+        $stubDateTime = new DateTimeImmutable("2018-02-13 18:50:00");
+        $inMemoryClient = new InMemoryClientAdapter(new ClockStub($stubDateTime));
+
+        // When
+        $inMemoryClient->histogram(
+            "histogram-stat",
+            23488,
+            0.55,
+            ["histogram" => "yep", "has-tags" => "also yes"]
+        );
+
+        // Then
+        $statsRecord = $inMemoryClient->getStats();
+        self::assertCount(1, $statsRecord->histogram);
+
+        $histogramRecord = $statsRecord->histogram[0];
+        self::assertEquals("histogram-stat", $histogramRecord->stat);
+        self::assertEquals(23488, $histogramRecord->value);
+        self::assertEquals(0.55, $histogramRecord->sampleRate);
+        self::assertEqualsCanonicalizing(["histogram" => "yep", "has-tags" => "also yes"], $histogramRecord->tags);
+        self::assertEquals($stubDateTime, $histogramRecord->recordedAt);
+    }
+
+    #[Test]
+    public function normalizesTags()
+    {
+        // Given
+        $inMemoryClient = new InMemoryClientAdapter(new ClockStub(new DateTimeImmutable()));
+
+        // And
+        $tagNormalizerSpy = new TagNormalizerSpy();
+        $inMemoryClient->setTagNormalizer($tagNormalizerSpy);
+
+        // When
+        $inMemoryClient->histogram(stat: "irrelevant", value: 19.2, tags: ["hello" => "world"]);
+
+        // Then
+        $this->assertEqualsCanonicalizing([["hello" => "world"]], $tagNormalizerSpy->getNormalizeCalls());
+    }
+}
